working on seed extension

+++ src/core/datastructure/keywordtreebyte/keywordtreebyte.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Position struct {
-	SequenceIndex uint32
+	SequenceIndex uint8
 	Position      uint32
 }
